\name{data_vignette}
\alias{epi_rg}
\alias{epi_ss}
\alias{epi_lrr}
\alias{epi_cna}
\alias{epi_seg}
\docType{data}
\title{
Example datasets for Epicopy.
}
\description{
Example datasets for Epicopy, including RGChannelSet, sample sheet,
    LRR, CNA, and segmentation results.
}
\usage{data("data_vignette")}
\value{
\describe{
    \item{\code{epi_rg}}{RGChannelSet}
    \item{\code{epi_ss}}{Sample sheet}
    \item{\code{epi_lrr}}{LRR object}
    \item{\code{epi_cna}}{CNA object}
    \item{\code{epi_seg}}{Segmentation result}
  }
}
\examples{
data(data_vignette)
}
\keyword{datasets}
