% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicopy.R
\name{epicopy}
\alias{epicopy}
\title{Run Epicopy.}
\usage{
epicopy(target_dir, output_dir = NULL, project_name = NULL,
  Normals = NULL, sampNames = NULL, QN = FALSE, Ref = "mode",
  mode.bw = 0.1, mode.method = "naive", normal.cnv = NULL,
  mean.center = TRUE, filterProbes = FALSE, retainedProbes = NULL,
  ncores = 1L, seg.undo.splits = "sdundo", seg.undo.SD = 2,
  filterbycount = TRUE, min_probes = 50, verbose = TRUE, ...)
}
\arguments{
\item{target_dir}{Target directory with IDAT files and sample sheet.}

\item{output_dir}{Output directory. Defaults to current directory. If FALSE,
no GISTIC 2.0 files will be outputted, but can be manually obtained
using the function \code{export_gistic}.}

\item{project_name}{Will be used as prefixes for any output.}

\item{Normals}{Character string identifying column in sample sheet to scan for
normals. Defaults to NULL, using all included Epicopy normals. See 
details for information on the normals and additional arguments.}

\item{sampNames}{Character string identifying column in sample sheet to use for
sample names. If NULL, uses the default \code{read.metharray.exp} values, which are
the ChipNo_Position.}

\item{QN}{Logical. Quantile normalize post funnorm? Defaults to \code{FALSE}.}

\item{Ref}{String. One of 'mode' or 'median'. How to calculate LRR.}

\item{mode.bw}{If Ref = 'mode', uses this bw to estimate the mode. 
Defaults to 0.1.}

\item{mode.method}{If Ref = 'mode', uses this method to estimate 
the mode. Defaults to 'naive'.}

\item{normal.cnv}{Logical. Defaults to NULL which evaluates the Normal input.
If Normal is user input normal tissues, then defaults to TRUE. Otherwise FALSE.}

\item{mean.center}{Logical. Perform mean centering of final data.}

\item{filterProbes}{Logical. Filter probe by TCGA defined 450K probes 
adjacent to a known SNP.}

\item{retainedProbes}{Character. IF filterProbes is TRUE, denotes probes to keep. 
If NULL, uses probes retained by TCGA.}

\item{ncores}{Integer. Number of cores to use for segmentation. Defaults to 1.}

\item{seg.undo.splits}{Undo.splits argument for \code{segment} from
\code{DNAcopy}. Default: \code{'sdundo'}.}

\item{seg.undo.SD}{undo.SD argument for \code{segment}. Default: 2.}

\item{filterbycount}{Logical. Recommended. Should the output segment file be 
filtered for having at least \code{min_probes} number of probes in the 
segment.}

\item{min_probes}{Number of probes to filter against.}

\item{verbose}{Logical. Verbose?}

\item{...}{Passes argument to \code{getLRR}.}
}
\value{
Epicopy object and output files.
}
\description{
Parses the CNA result from \code{LRRtoCNA} and writes a
   segment file and marker file for input into GISTIC 2.0.
}
\details{
\code{epicopy} Master function to run epicopy from raw Illumina Methylation 450K data.

Epicopy reads the sample sheet provided by the user in the \code{target_dir},
imports the experiment, and returns a CNA file while writing gistic outputs at
the \code{output_dir}.

If the user has their own normals, specify the column name in the samplesheet
that contains the identifier 'normal' (not case sensitive). Else, they can
specify one of 'all' (default, NULL), 'breast', 'lung', or 'thyroid' to use Epicopy 
included normals. The last alternative is NA which uses all the samples as 
references, under the assumption that the \code{mode}/\code{median} of all the
samples are at 2n. The last is recommended either in samples with low CNV, or
when there are many samples. See also \code{\link[Epicopy]{getLRR}}.
}
\examples{

\dontrun{
# Run epicopy on included example TCGA data
input_loc <- system.file('extdata', 'raw_idat', package = 'Epicopy')
# No output directory, returns only local R object
epi_seg <- epicopy(input_loc, output_dir = FALSE) 
}

}
\seealso{
\code{\link[Epicopy]{getLRR}} \code{\link[Epicopy]{LRRtoCNA}} 
   \code{\link[Epicopy]{export_gistic}} \code{\link[minfi]{RGChannelSet-class}}
   \code{\link[minfi]{read.metharray.sheet}} \code{\link[minfi]{read.metharray.exp}}
}
