% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRR.R
\name{getLRR}
\alias{getLRR}
\title{Obtain LRR from RGset}
\usage{
getLRR(rgSet, Normals = NULL, sampNames = NULL, QN = FALSE,
  Ref = "mode", mode.bw = 0.1, mode.method = "naive", normal.cnv = TRUE,
  mean.center = TRUE, filterProbes = FALSE, retainedProbes = NULL,
  keep_fnobj = FALSE, fn_output = NULL, ...)
}
\arguments{
\item{rgSet}{RGChannelSet object containing samples of interest}

\item{Normals}{Either (1) RGChannelSet of normals (reference set), (2) a 
logical or numeric vector  denoting the normals in the input rgSet, (3) a
character of 'thyroid', 'lung', 'breast', or 'all' denoting reference 
set included in Epicopy, or (4) defaults to NULL, which uses all normals 
in Epicopy.}

\item{sampNames}{A character vector giving sample names. Defaults to NULL, 
which uses \code{pData(rgSet)$Sample_Name}.}

\item{QN}{Logical. Perform probe type specific quantile normalization of 
reference set.}

\item{Ref}{'mode' or 'median'. Method to calculate reference intensity. 
Defaults to mode using the modeest package.}

\item{mode.bw}{If Ref = 'mode', uses this bw to estimate the mode. 
Defaults to 0.1.}

\item{mode.method}{If Ref = 'mode', uses this method to estimate 
the mode. Defaults to 'naive'.}

\item{normal.cnv}{Logical. Return LRR for reference set.}

\item{mean.center}{Logical. Perform mean centering of final data.}

\item{filterProbes}{Logical. Filter probe by TCGA defined 450K probes 
adjacent to a known SNP.}

\item{retainedProbes}{Character. IF filterProbes is TRUE, denotes probes to keep. 
If NULL, uses probes retained by TCGA.}

\item{keep_fnobj}{Logical. Save functional normalized object as rda.}

\item{fn_output}{Directory for fnobject. Defaults to current directory.}

\item{...}{Passes argument to preprocessFunnorm.}
}
\value{
Matrix containing LRR values.
}
\description{
Returns a matrix containing the log R ratios (LRR) of samples 
   to reference set. Input file has to be an RGset. Normals can be defined 
   by the user (see details).
}
\details{
\code{getLRR} Preprocess RGset and obtain LRR for samples against a 
   reference set.

GetLRR takes RGset as an argument as an input.

Normals/Reference samples can be defined multiple ways. Defaults to 
NULL which uses all the normals incorporated by the Epicopy package 
(see next paragraph). Providing an RGset will BiocGenerics::combine(rgSet, Normals) to 
get a dataset. Providing a logical/numeric vector will subset the rgSet 
into query and reference, with TRUE/numbers representing reference samples. 
Users can also specify one of 'thyroid', 'breast', 'lung', or 'all' to use 
Epicopy normals.

Epicopy normals are compiled using solid normal tissue methylation microarrays 
from TCGA.
}
\examples{

# Get LRR
data(data_vignette)
# Median is used for quick calculation
res_lrr <- getLRR(epi_rg, Normals = NA, Ref = 'median')

}
\seealso{
\code{\link[minfi]{preprocessFunnorm}} 
   \code{\link[minfi]{RGChannelSet-class}}
}
