% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/LRRtoCNA.R
\name{LRRtoCNA}
\alias{LRRtoCNA}
\title{Convert LRR to segmented copy number}
\usage{
LRRtoCNA(LRR, sampID = NULL, ncores = 1L, seg.undo.splits = "sdundo",
  seg.undo.SD = 2, ...)
}
\arguments{
\item{LRR}{Matrix of LRR from \code{getLRR}}

\item{sampID}{Sample IDs. Default: NULL, uses \code{colnames(LRR)}}

\item{ncores}{Integer. Default: 1L. Define number of cores for parallel
segmentation. See details.}

\item{seg.undo.splits}{Undo.splits argument for \code{segment} from
\code{DNAcopy}. Default: \code{'sdundo'}.}

\item{seg.undo.SD}{undo.SD argument for \code{segment}. Default: 2.}

\item{...}{Passes arguments to \code{segment}.}
}
\value{
A segmented DNAcopy object. See \code{\link[DNAcopy]{segment}}.
}
\description{
\code{LRRtoCNA} Converts LRR to segmented copy number using the circular
   binart segmentation (CBS) algorithm.
}
\details{
This function coerces the LRR file into a suitable input for
   the \code{CNA}, \code{smooth.CNA}, and \code{segment} algorithm
   provided by the \code{DNAcopy} package. Parallelization for segmentation
   of \code{ParDNAcopy} package is implemented when \code{ncores} >= 2L.
}
\examples{
# Load example data
data(data_vignette)
# Run segmentation
epi_cna <- LRRtoCNA(epi_lrr, seg.undo.splits = 'sdundo', seg.undo.SD = 2)
}
\seealso{
\code{\link[DNAcopy]{CNA}}
   \code{\link[DNAcopy]{smooth.CNA}}
   \code{\link[DNAcopy]{segment}}
   \code{\link[ParDNAcopy]{parSegment}}
}

