% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_segments.R
\name{plot_segments}
\alias{plot_segments}
\title{Plot segmented files.}
\usage{
plot_segments(CNA, which_sample, threshold = 0.15, ylab = "Copy Number",
  xlab = "Genomic Location (kb)", ylim = c(-2, 2), title = NULL,
  seg_lwd = 2, seg_col = c("skyblue3", "black"), ...)
}
\arguments{
\item{CNA}{CNA output from LRRtoCNA}

\item{which_sample}{Either a numeric/integer to indicate which sample to plot or
a character vector with the sample name.}

\item{threshold}{Numeric vector indicating where to plot dotted lines to indicate
threshold. Defaults at 0.15.}

\item{ylab}{Y-axis labels}

\item{xlab}{X-axis labels}

\item{ylim}{Y-axis limits}

\item{title}{Title. Defaults to sample name with Epicopy segmentation string.}

\item{seg_lwd}{Line width of the segments}

\item{...}{Passes arguments to plot to control everything other than segment.}
}
\value{
Segment plots.
}
\description{
Plots segmentation output for individual samples.
}
\details{
\code{epicopy} Plot resulting segments from \code{LRRtoCNA}.

Plots the segmentation results from \code{LRRtoCNA}.
}
\examples{

# Load included example data.
data(data_vignette)
# Plot
plot_segments(epi_seg, which_sample = 1)

}
\seealso{
\code{\link[Epicopy]{getLRR}} \code{\link[Epicopy]{LRRtoCNA}} 
   \code{\link[Epicopy]{export_gistic}} \code{\link[minfi]{RGChannelSet-class}}
   \code{\link[minfi]{read.450k.sheet}} \code{\link[minfi]{read.450k.exp}}
}
