% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotCoverage.R
\name{plotCoverage}
\alias{plotCoverage}
\title{Plot coverage of 450K across hg19}
\usage{
plotCoverage(chr, winSize = 1e+06, maxVal = 50, chrRange = NULL, ...)
}
\arguments{
\item{chr}{Chromosome to plot.}

\item{winSize}{Window size of segments.}

\item{maxVal}{Upper limit of number of probes for darkest color.}

\item{chrRange}{Region (in bases) of the chromosome to plot. If NULL plots
   the whole chromosome.}
}
\description{
Plots the coverage of 450K array probes of specified chromosome
   range in specified window sizes.
}
\details{
\code{plotCoverage} Plots the coverage of 450K array across the genome.

Uses hg19 annotations. \code{winSize} specifies window sizes and
   \code{maxVal} controls the number of probes for the darkest color in the
   window.
}
\examples{
# Plot coverage through the whole of chromosome two
plotCoverage('chr2')

# Plot coverage through the whole of chromosome two with windows of 2Mb and max of 200 probes
plotCoverage('chr2', winSize = 2e6, maxVal = 200)
}

